# ci: speedup julia test & setup ci for pr