The startup and pre-compilation costs of julia are high,
so for simple computations you can skip the pre-compilation and optimization.

```
Baseline:
    santity:    88s
    benchmarks: 604s
    binary32:   724s
    binary64:   721s

Pr:
    santity:    29s
    benchmarks: 216s
    binary32:   269s
    binary64:   301s
```