@inkydragon 2024-01-30
It looks like no CI will be run for pr.
I would suggest running sanity checks only for pr and running full tests only on the master branch.