@inkydragon 2024-01-30
I dropped the commit that modified the Github Action.
It looks like the CI doesn't trigger for this pr though.